fix: react query 설정을 preview 로 이동
